<fix>[network]: reformat L2NetworkHostRefVO
1. add bridgeName field to L2NetworkHostRefVO to support different
bridges when realizing l2 on different hosts;
2. remove the attachStatus field from L2NetworkHostRefVO and keep the
logic consistent with UplinkGroupVO;

DBImpact

Resolves: ZSV-6396

+++ network/src/main/java/org/zstack/network/l2/L2NetworkHostUtils.java
@@ -7,73 +7,79 @@ import org.zstack.header.errorcode.ErrorCode;
 import org.zstack.header.host.HostParam;
 import org.zstack.header.host.HostVO;
 import org.zstack.header.host.HostVO_;
-import org.zstack.header.network.l2.L2NetworkAttachStatus;
 import org.zstack.header.network.l2.L2NetworkHostRefVO;
 import org.zstack.header.network.l2.L2NetworkHostRefVO_;
+import org.zstack.header.network.l2.L2ProviderType;
 import org.zstack.utils.CollectionUtils;
 import org.zstack.utils.Utils;
 import org.zstack.utils.logging.CLogger;
 
 import java.util.ArrayList;
+import java.util.Collections;
 import java.util.List;
+import java.util.stream.Collectors;
 
 import static org.zstack.core.Platform.argerr;
 
 public class L2NetworkHostUtils {
     private static final CLogger logger = Utils.getLogger(L2NetworkHostUtils.class);
 
-    public static void deleteL2NetworkHostRef(String l2NetworkUuid, List<String> hostUuids) {
-        logger.debug(String.format("del L2NetworkHostRefVO, l2NetworkUuid:%s, hostUuids:%s",
-                l2NetworkUuid, hostUuids));
+    public static void deleteL2NetworkHostRef(String l2Uuid, List<String> hostUuids) {
+        if (CollectionUtils.isEmpty(hostUuids)) {
+            return;
+        }
+
+        logger.debug(String.format("del L2NetworkHostRefVO, l2NetworkUuid: %s, hostUuids: %s",
+                l2Uuid, hostUuids));
         SQL.New(L2NetworkHostRefVO.class)
-                .eq(L2NetworkHostRefVO_.l2NetworkUuid, l2NetworkUuid)
+                .eq(L2NetworkHostRefVO_.l2NetworkUuid, l2Uuid)
                 .in(L2NetworkHostRefVO_.hostUuid, hostUuids)
                 .delete();
     }
 
-    public static void changeL2NetworkToHostRefDetached(String l2NetworkUuid, String hostUuid) {
-        if (checkIfL2NetworkHostRefNotExist(l2NetworkUuid, hostUuid)) {
-            logger.warn(String.format("can not find host l2 network ref[l2NetworkUuid: %s, hostUuid: %s]",
-                    l2NetworkUuid, hostUuid));
+    public static void deleteL2NetworkHostRef(String l2Uuid, String hostUuid) {
+        deleteL2NetworkHostRef(l2Uuid, Collections.singletonList(hostUuid));
+    }
+
+    public static void deleteL2NetworkHostRef(List<String> l2Uuids, String hostUuid) {
+        if (CollectionUtils.isEmpty(l2Uuids)) {
             return;
         }
 
-        logger.debug(String.format("change L2NetworkHostRefVO to Detached, l2NetworkUuid:%s, hostUuid:%s",
-                l2NetworkUuid, hostUuid));
+        logger.debug(String.format("del L2NetworkHostRefVO, l2NetworkUuids: %s, hostUuid: %s",
+                l2Uuids, hostUuid));
         SQL.New(L2NetworkHostRefVO.class)
+                .in(L2NetworkHostRefVO_.l2NetworkUuid, l2Uuids)
                 .eq(L2NetworkHostRefVO_.hostUuid, hostUuid)
-                .eq(L2NetworkHostRefVO_.l2NetworkUuid, l2NetworkUuid)
-                .set(L2NetworkHostRefVO_.attachStatus, L2NetworkAttachStatus.Detached).update();
+                .delete();
     }
 
-    public static void changeL2NetworkToHostRefAttached(String l2NetworkUuid, String hostUuid) {
-        if (checkIfL2NetworkHostRefNotExist(l2NetworkUuid, hostUuid)) {
-            logger.warn(String.format("can not find host l2 network ref[l2NetworkUuid: %s, hostUuid: %s]",
-                    l2NetworkUuid, hostUuid));
-            return;
-        }
-
-        logger.debug(String.format("change L2NetworkHostRefVO to Attached, l2NetworkUuid:%s, hostUuid:%s",
-                l2NetworkUuid, hostUuid));
-        SQL.New(L2NetworkHostRefVO.class)
-                .eq(L2NetworkHostRefVO_.hostUuid, hostUuid)
-                .eq(L2NetworkHostRefVO_.l2NetworkUuid, l2NetworkUuid)
-                .set(L2NetworkHostRefVO_.attachStatus, L2NetworkAttachStatus.Attached).update();
-    }
-
-    public static boolean checkIfL2NetworkHostRefNotExist(String l2NetworkUuid, String hostUuid) {
-        return !Q.New(L2NetworkHostRefVO.class)
-                .eq(L2NetworkHostRefVO_.l2NetworkUuid, l2NetworkUuid)
-                .eq(L2NetworkHostRefVO_.hostUuid, hostUuid)
-                .isExists();
-    }
-
-    public static boolean checkIfL2AttachedToHost(String l2NetworkUuid, String hostUuid) {
+    public static String getBridgeNameFromL2NetworkHostRef(String l2Uuid, String hostUuid) {
         return Q.New(L2NetworkHostRefVO.class)
-                .eq(L2NetworkHostRefVO_.l2NetworkUuid, l2NetworkUuid)
+                .select(L2NetworkHostRefVO_.bridgeName)
+                .eq(L2NetworkHostRefVO_.l2NetworkUuid, l2Uuid)
                 .eq(L2NetworkHostRefVO_.hostUuid, hostUuid)
-                .eq(L2NetworkHostRefVO_.attachStatus, L2NetworkAttachStatus.Attached)
+                .findValue();
+    }
+
+    public static boolean checkIfL2AttachedToHost(String l2Uuid, String hostUuid) {
+        return Q.New(L2NetworkHostRefVO.class)
+                .eq(L2NetworkHostRefVO_.l2NetworkUuid, l2Uuid)
+                .eq(L2NetworkHostRefVO_.hostUuid, hostUuid)
                 .isExists();
+    }
+
+    public static List<String> getExcludeHostUuids(List<String> l2Uuids, List<String> hostUuids) {
+        if (CollectionUtils.isEmpty(l2Uuids) || CollectionUtils.isEmpty(hostUuids)) {
+            return new ArrayList<>();
+        }
+
+        List<String> hostsWithRef = Q.New(L2NetworkHostRefVO.class)
+                .select(L2NetworkHostRefVO_.hostUuid)
+                .in(L2NetworkHostRefVO_.l2NetworkUuid, l2Uuids)
+                .in(L2NetworkHostRefVO_.hostUuid, hostUuids)
+                .listValues();
+        return hostUuids.stream().filter(it -> !hostsWithRef.contains(it)).collect(Collectors.toList());
     }
 
     public static ErrorCode validateHostParams(List<HostParam> hostParams, String clusterUuid, String hostUuid) {
@@ -107,6 +113,11 @@ public class L2NetworkHostUtils {
 
             if (StringUtils.isEmpty(hostParam.getPhysicalInterface())) {
                 return argerr("physical interface can not be null in HostParam");
+            }
+
+            if (!StringUtils.isEmpty(hostParam.getL2ProviderType())
+                    && !L2ProviderType.hasType(hostParam.getL2ProviderType())) {
+                return argerr("unsupported l2Network provider type[%s]", hostParam.getL2ProviderType());
             }
         }
